// third-party/polymarket/src/gen/polymarket/polymarket.ts
export enum Side {
  SIDE_UNSPECIFIED = 0,
  SIDE_BUY = 1,
  SIDE_SELL = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum AssetType {
  ASSET_TYPE_UNSPECIFIED = 0,
  ASSET_TYPE_COLLATERAL = 1,
  ASSET_TYPE_CONDITIONAL = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum PriceHistoryInterval {
  PRICE_HISTORY_INTERVAL_UNSPECIFIED = 0,
  PRICE_HISTORY_INTERVAL_MAX = 1,
  PRICE_HISTORY_INTERVAL_ONE_WEEK = 2,
  PRICE_HISTORY_INTERVAL_ONE_DAY = 3,
  PRICE_HISTORY_INTERVAL_SIX_HOURS = 4,
  PRICE_HISTORY_INTERVAL_ONE_HOUR = 5,
  UNRECOGNIZED = -1,
}
// ...
export type PolymarketServiceService = typeof PolymarketServiceService;